+++ crates/fiber-json-types/src/serde_utils.rs
@@ -336,3 +336,111 @@ pub mod duration_hex {
         Ok(Duration::from_secs(seconds))
     }
 }
+
+/// Convert SCREAMING_SNAKE_CASE to PascalCase.
+/// Example: "OUR_INIT_SENT" -> "OurInitSent"
+pub(crate) fn to_pascal_case(s: &str) -> String {
+    s.split('_')
+        .map(|word| {
+            let mut chars = word.chars();
+            match chars.next() {
+                None => String::new(),
+                Some(first) => {
+                    first.to_uppercase().collect::<String>() + &chars.as_str().to_lowercase()
+                }
+            }
+        })
+        .collect()
+}
+
+/// Convert PascalCase to SCREAMING_SNAKE_CASE.
+/// Example: "OurInitSent" -> "OUR_INIT_SENT"
+pub(crate) fn to_snake_case(s: &str) -> String {
+    let mut result = String::new();
+    for (i, c) in s.chars().enumerate() {
+        if c.is_uppercase() && i > 0 {
+            result.push('_');
+        }
+        result.push(c.to_ascii_uppercase());
+    }
+    result
+}
+
+/// Macro to define flags types that serialize to PascalCase strings.
+/// For single flag, returns the flag name in PascalCase (e.g., "OurInitSent").
+/// For multiple flags, returns comma-separated names (e.g., "OurInitSent,TheirInitSent").
+#[macro_export]
+macro_rules! define_rpc_flags {
+    (
+        $(#[$struct_meta:meta])*
+        pub struct $name:ident($ty:ty) {
+            $($(#[$flag_meta:meta])* const $flag_name:ident = $flag_value:expr;)*
+        }
+    ) => {
+        $(#[$struct_meta])*
+        pub struct $name(pub $ty);
+
+        impl $name {
+            $(pub const $flag_name: $ty = $flag_value;)*
+
+            #[allow(clippy::wrong_self_convention)]
+            fn to_strings(self) -> Vec<String> {
+                let mut names = Vec::new();
+                $(
+                    if self.0 & Self::$flag_name != 0 {
+                        names.push($crate::serde_utils::to_pascal_case(stringify!($flag_name)));
+                    }
+                )*
+                names
+            }
+
+            fn from_string(s: &str) -> Option<Self> {
+                let mut flags: $ty = 0;
+                for name in s.split(',') {
+                    let name = name.trim();
+                    let snake_name = $crate::serde_utils::to_snake_case(name);
+                    match snake_name.as_str() {
+                        $(stringify!($flag_name) => flags |= Self::$flag_name,)*
+                        _ => return None,
+                    }
+                }
+                Some($name(flags))
+            }
+        }
+
+        impl serde::Serialize for $name {
+            fn serialize<S: serde::Serializer>(&self, serializer: S) -> Result<S::Ok, S::Error> {
+                let names = self.clone().to_strings();
+                if names.is_empty() {
+                    serializer.serialize_str("")
+                } else {
+                    serializer.serialize_str(&names.join(","))
+                }
+            }
+        }
+
+        impl<'de> serde::Deserialize<'de> for $name {
+            fn deserialize<D: serde::Deserializer<'de>>(deserializer: D) -> Result<Self, D::Error> {
+                let s = String::deserialize(deserializer)?;
+                if s.is_empty() {
+                    Ok($name(0))
+                } else {
+                    $name::from_string(&s)
+                        .ok_or_else(|| serde::de::Error::custom(format!("Invalid {}: {}", stringify!($name), s)))
+                }
+            }
+        }
+
+        impl From<$ty> for $name {
+            fn from(v: $ty) -> Self {
+                $name(v)
+            }
+        }
+
+        impl From<$name> for $ty {
+            fn from(v: $name) -> Self {
+                v.0
+            }
+        }
+    };
+}
